% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeason.R
\name{getSeason}
\alias{getSeason}
\title{Get season from date}
\usage{
getSeason(x, seasonlab = c("Winter", "Spring", "Summer", "Fall"))
}
\arguments{
\item{x}{a POSIXct, POSIXlt, Date, chron, yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, or fts object..}

\item{seasonlab}{Season labels, concatenated strings with 4 seasons, default c("Winter","Spring","Summer","Fall")}
}
\value{
Returns the corresponding season.
}
\description{
Get the season name out of a date
}
\examples{
x <- as.Date("2019-04-01")
getSeason(x) #"Spring"

x <- as.Date("2019-04-01")
getSeason(x, seasonlab = c("Winter","Frühling","Sommer","Herbst")) #"Frühling"
}
