% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTypEleConsHousehold.R
\name{getTypEleConsHousehold}
\alias{getTypEleConsHousehold}
\title{Get typical electricity consumption of a household}
\usage{
getTypEleConsHousehold(
  occupants = 3,
  rooms = NULL,
  bldgType = "single",
  dishwasher = "classic",
  freezer = "classic",
  cookingBaking = "normal",
  effLighting = "mix",
  dryer = "classic",
  laundry = "classic",
  waterHeater = "none",
  eleCommon = "excluded"
)
}
\arguments{
\item{occupants}{numeric number of occupants, minimum 1, maximum 6}

\item{rooms}{numeric number of rooms, minimum 1, maximum 9, defaul NULL and internally assumed by the number of occupants}

\item{bldgType}{type of building, single family house or multi dwelling untit (\code{"single"} or \code{"multi"}, default \code{"multi"})}

\item{dishwasher}{type of dishwasher (\code{"none"}, \code{"classic"} or \code{"hotWaterSupply"}, default \code{"classic"})}

\item{freezer}{type of freezer (\code{"none"} or \code{"classic"}, default \code{"classic"})}

\item{cookingBaking}{cooking and baking behaviour (\code{"occasionally"}, \code{"normal"} or \code{"intensive"}, default \code{"normal"})}

\item{effLighting}{energy efficient lighting (\code{"minority"}, \code{"mix"} or \code{"majority"}, default \code{"mix"})}

\item{dryer}{type of clothing dryer (\code{"none"}, \code{"roomAir"}, \code{"heatPump"} or \code{"classic"}, default \code{"classic"})}

\item{laundry}{type of laundry machine (\code{"none"}, \code{"classic"} or \code{"hotWaterSupply"}, default \code{"classic"})}

\item{waterHeater}{type of water heater (\code{"none"}, \code{"electric"} or \code{"heatpump"}, default \code{"none"})}

\item{eleCommon}{common electricity (\code{"included"} or \code{"excluded"}, default \code{"excluded"})}
}
\value{
Returns a typical electricity consumtion of a comparable Swiss household in kWh/year.
}
\description{
Get a typical electricity consumption of a household in kWh/year. This is useful to compare a real dataset with a typical consumption value.
}
\details{
Source:
Nipkov, J. (2013). \href{https://www.aramis.admin.ch/Default.aspx?DocumentID=61764}{Typischer Haushalt-Stromverbrauch - Schlussbericht}. Bundesamt für Energie (BFE).

Occupants:
\itemize{
\item The persons should be present at least on working days, i.e. at least in the evening and usually for 2 meals.
\item Children up to approx. 10 years are to be counted as approx. "½ person".
\item Teenagers from about 11 years on are to be counted as adults.
}

Rooms:
\itemize{
\item If the effective room number deviates significantly, i.e. by more than 1, from a "typical room number", the typical energy consumption gets adjusted accordingly.
}

Dishwasher:
\itemize{
\item none: no dishwasher powered with this meter
\item classic: dishwasher available, no further classification
\item with hot water supply: electrical dishwasher which uses as well hot water to heat up
}

Freezer:
\itemize{
\item none: no "big" freezer powered with this meter. The one in the refrigerator does not count.
\item classic: additional freezer available, no further classification
}

Cooking & Baking:
\itemize{
\item If cooking and baking is particularly intensive and often, and little hot food is eaten outdoors, this increases the base value by 50 per person.
\item Conversely, if the stove and oven are used less intensively and food is eaten outdoors frequently, this reduces the base value by 50 per person.
}

Efficient Lighting:
\itemize{
\item The mix values apply to a share of efficient lamps of about 40 to 50\% (energy saving lamps, LED, fluorescent tubes).
\item Attention: Halogen or "Eco-Halogen" lamps are not efficient!
\item Majority means more than 80\%, minority less than 20\%
}

Clothes Dryer:
\itemize{
\item none: no dryer powered with this meter
\item room air dryer: room air dehumidifier in the drying room
\item heat pumpt dryer: dryer with integrated heat pump, uses less energy than classic one!
\item classic dryer: classical one which is only connected to the electric power
}

Washing Machine:
\itemize{
\item none: no washing machine powered with this meter
\item classic: washing machine available, no further classification
\item with hot water supply: electrical washing machine which uses as well hot water to heat up
}

Water Heater:
\itemize{
\item none: no water heater powered with this meter
\item Electric Boiler: separate electric boiler powered with this meter
\item Heat Pump: only for single family houses when the water heater is connected to the meter
}

Common electricity (building equipment, common lighting):
\itemize{
\item included: e.g. for single family houses with main meter
\item excluded:  e.g. for multi family houses where the meter of the flat gets analyzed. Normally the common electricity is measured by a separate meter.
}
}
\examples{
# single family house
getTypEleConsHousehold(occupants=3, rooms=5.5, bldgType="single", waterHeater="heatpump", eleCommon="included")

# flat in a multi family house
getTypEleConsHousehold(occupants=3, rooms=4.5, bldgType="multi", freezer="none", eleCommon="excluded")
}
