% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mollierHxDiagram.R
\name{mollierHxDiagram}
\alias{mollierHxDiagram}
\title{D3 Mollier hx Diagram}
\usage{
mollierHxDiagram(
  data,
  graphTempMin = 15,
  graphTempMax = 30,
  graphHumAbsMin = 0,
  graphHumAbsMax = 0.017,
  altitude = 450,
  cmfZoneTempMin = 21,
  cmfZoneTempMax = 26,
  cmfZoneHumRelMin = 30,
  cmfZoneHumRelMax = 50,
  cmfZoneHumAbsMin = 0,
  cmfZoneHumAbsMax = 0.012
)
}
\arguments{
\item{data}{Dataset to use for scatter plot. Must be a data.frame with "datetime, temperature °C, humidity \%rH"}

\item{graphTempMin}{Graph temperature minimum on y-axis in °C, default 15.0}

\item{graphTempMax}{Graph temperature maximum on y-axis in °C, default 30.0}

\item{graphHumAbsMin}{Graph humidity minimum on x-axis in kg/kg, default 0.0}

\item{graphHumAbsMax}{Graph humidity maximun on x-axis in kg/kg, default 0.017}

\item{altitude}{Altitude in meters above sea level to calculate pressure, default 450}

\item{cmfZoneTempMin}{Comfort zone temperature minimum in °C, default 21.0}

\item{cmfZoneTempMax}{Comfort zone temperature maximum in °C, default 26.0}

\item{cmfZoneHumRelMin}{Comfort zone humidity minimum in \%rH, default 30.0}

\item{cmfZoneHumRelMax}{Comfort zone humidity maximum in \%rH, default 50.0}

\item{cmfZoneHumAbsMin}{Comfort zone humidity minimum in kg/kg, default 0.0}

\item{cmfZoneHumAbsMax}{Comfort zone humidity maximum in kg/kg, default 0.012. Also called sultriness limit (de: Schwülgrenze)}
}
\value{

}
\description{
Creates D3 Mollier hx Diagram with scatter plot and comfort zone
}
\examples{
library(redutils)
library(r2d3)
library(dplyr)
data <- readRDS(system.file("sampleData/flatTempHum.rds", package = "redutils"))
mollierHxDiagram(data)

}
