% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDailyProfilesDecomposed.R
\name{plotDailyProfilesDecomposed}
\alias{plotDailyProfilesDecomposed}
\title{Plot Daily Profiles Decomposed}
\usage{
plotDailyProfilesDecomposed(
  data,
  locTimeZone = "UTC",
  main = "Daily Profiles - Decomposed",
  ylab = "delta Energy Consumption",
  k = 672
)
}
\arguments{
\item{data}{Dataset to use for plot, minimum 1 hour aggregated. Must be a data.frame with "timestamp YmdHMS, energy consumption"}

\item{locTimeZone}{Time zone of timestamp, default "UTC"}

\item{main}{Main title of plot, default "Daily Profiles - Decomposed"}

\item{ylab}{y-axis title, default "delta Energy Consumption"}

\item{k}{Amount of samples for the rolling window used for the decomosition of the trend component, default 672 which corresponds to 7 days}
}
\value{
Returns a ggplot object
}
\description{
Plot a Graph with Decomposed Daily Energy Consumption Profiles by Weekday
}
\examples{
data <- readRDS(system.file("sampleData/eboBookEleMeter.rds", package = "redutils"))
plotDailyProfilesDecomposed(data, locTimeZone = "Europe/Zurich")
}
