% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnergyConsBeforeAfter.R
\name{plotEnergyConsBeforeAfter}
\alias{plotEnergyConsBeforeAfter}
\title{Plot Energy Consumption Before/After Optimization}
\usage{
plotEnergyConsBeforeAfter(
  data,
  dateOptimization,
  locTimeZone = "UTC",
  titlePlot = "Before/After Optimization",
  titleYAxis = "Energy Consumption \\n(kWh/month)"
)
}
\arguments{
\item{data}{Dataset to use for plot with montly aggregated data. Must be a data.frame with "timestamp, energy consumption"}

\item{dateOptimization}{Date of Optimization in format YYYY-MM-DD}

\item{locTimeZone}{Time zone of timestamp, default "UTC"}

\item{titlePlot}{Main title of plot, default "Before/After Optimization"}

\item{titleYAxis}{y-axis title, default "Energy Consumption (kWh/month)"}
}
\value{
Returns a ggplot object
}
\description{
Plot a Graph with Energy Consumption per Month before/after an Optimization
}
\examples{
data <- readRDS(system.file("sampleData/flatHeatingEnergy.rds", package = "redutils"))
plotEnergyConsBeforeAfter(data, dateOptimization = "2017-09-01")
}
