% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnergyConsDailyProfileOverview.R
\name{plotEnergyConsDailyProfileOverview}
\alias{plotEnergyConsDailyProfileOverview}
\title{Plot Energy Consumption Daily Profile Overview}
\usage{
plotEnergyConsDailyProfileOverview(
  data,
  locTimeZone = "UTC",
  titlePlot = "Daily Energy Consumption by Weekday and Season",
  titleYAxis = "Energy Consumption (kWh/h)"
)
}
\arguments{
\item{data}{Dataset to use for plot, minimum 1 hour aggregated. Must be a data.frame with "timestamp YmdHMS, energy consumption"}

\item{locTimeZone}{Time zone of timestamp, default "UTC"}

\item{titlePlot}{Main title of plot, default "Before/After Optimization"}

\item{titleYAxis}{y-axis title, default "Energy Consumption (kWh/month)"}
}
\value{
Returns a ggplot object
}
\description{
Plot a Graph with Daily Energy Consumption Profiles by Weekday and Season
}
\examples{
data <- readRDS(system.file("sampleData/eboBookEleMeter.rds", package = "redutils"))
plotEnergyConsDailyProfileOverview(data, locTimeZone = "Europe/Zurich")
}
