% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmapMedian.R
\name{plotHeatmapMedianWeeks}
\alias{plotHeatmapMedianWeeks}
\title{Plot Heatmap Median Weeks}
\usage{
plotHeatmapMedianWeeks(
  data,
  locTimeZone = "UTC",
  main = "Heatmap Median per hour by Weekday and Season",
  ylab = "Energy Consumption (kWh/h)"
)
}
\arguments{
\item{data}{Dataset to use for plot, minimum 1 hour aggregated. Must be a data.frame with "timestamp YmdHMS, energy consumption"}

\item{locTimeZone}{Time zone of timestamp, default "UTC"}

\item{main}{Main title of plot, default "Heatmap Median per hour by Weekday and Season"}

\item{ylab}{y-axis title, default "Energy Consumption (kWh/month)"}
}
\value{
Returns a ggplot object
}
\description{
Plot Heatmap of Median Energy Consumption by Hour, Weekdays and Seasons of Year
}
\examples{
data <- readRDS(system.file("sampleData/eboBookEleMeter.rds", package = "redutils"))
plotHeatmapMedianWeeks(data, locTimeZone = "Europe/Zurich")
}
