% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeasonalXYBeforeAfter.R
\name{plotSeasonalXYBeforeAfter}
\alias{plotSeasonalXYBeforeAfter}
\title{Plot Seasonal Graph X/Y Before/After Optimization}
\usage{
plotSeasonalXYBeforeAfter(
  data,
  dateOptimization,
  locTimeZone = "UTC",
  main = "Before/After Optimization",
  ylab = "Energy Consumption \\n(kWh/month)"
)
}
\arguments{
\item{data}{Dataset to use for plot with montly aggregated data. Must be a data.frame with "timestamp, energy consumption"}

\item{dateOptimization}{Date of Optimization in format YYYY-MM-DD}

\item{locTimeZone}{Time zone of timestamp, default "UTC"}

\item{main}{Main title of plot, default "Before/After Optimization"}

\item{ylab}{y-axis title, default "Energy Consumption (kWh/month)"}
}
\value{
Returns a ggplot object
}
\description{
Plot a Seasonal Graph with x/y Axis per Month before/after an Optimization
}
\examples{
data <- readRDS(system.file("sampleData/flatHeatingEnergy.rds", package = "redutils"))
plotSeasonalXYBeforeAfter(data, dateOptimization = "2017-09-01")
}
