% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{season}
\alias{season}
\title{Get season from date}
\usage{
season(x, seasons = c("Winter", "Spring", "Summer", "Fall"))
}
\arguments{
\item{x}{a POSIXct, POSIXlt, Date, chron, yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, or fts object..}

\item{seasons}{concatenated strings with 4 seasons, default c("Winter","Spring","Summer","Fall")}
}
\value{
Returns the corresponding season.
}
\description{
Get the season name out of a date
}
\examples{
x <- as.Date("2019-04-01")
season(x) #"Spring"

x <- as.Date("2019-04-01")
season(x, c("Winter","Frühling","Sommer","Herbst")) #"Frühling"

}
